requirements.txt
asgiref==3.6.0
awsebcli==3.20.3
boto3
botocore
Django==4.1.5
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-storages
mysqlclient==2.1.1
psycopg2-binary==2.9.5
python-dotenv==0.21.0
pytz==2022.7
sendgrid==6.11.0
sqlparse==0.4.3
PyYAML==5.3.1
PyJWT==2.3.0